Renamed the app 'timetable_selector' to be 'view_timetables'
