+++ scripts/json_to_csv.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 
 # io settings
-loc = Path(__file__).parent / "timetable" / "timetable_selector" / "fixtures"
+loc = Path(__file__).parent / "timetable" / "view_timetables" / "fixtures"
 output_loc = Path(__file__).parent / "timetable" / "data_upload" / "test_data"
 files = ["classes"]
