+++ timetable/data/models/unsolved_class.py
@@ -30,7 +30,7 @@ class UnsolvedClass(models.Model):
     """
     Model used to specify the school_id classes that must take place, including who must be able to attend them,
     and also teaching hours / min number of slots etc. "Unsolved" since it represents an input to the solver which
-    finds the timetable structure that works across the board. Twin to "FixedClass" in timetable_selector app.
+    finds the timetable structure that works across the board. Twin to "FixedClass" in view_timetables app.
     """
     class_id = models.CharField(max_length=20, primary_key=True)
     school = models.ForeignKey(School, on_delete=models.CASCADE)
