+++ timetable/interfaces/view_timetables/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TimetableSelectorConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'interfaces.timetable_selector'
+    name = 'interfaces.view_timetables'
