+++ timetable/interfaces/view_timetables/static/javascript/collapsible.js
