+++ timetable/interfaces/view_timetables/static/javascript/table_to_csv.js
