+++ timetable/interfaces/view_timetables/static/stylesheets/pupil_teacher_navigators.css
