+++ timetable/interfaces/view_timetables/templates/selection_dashboard.html
