+++ timetable/interfaces/view_timetables/utils.py
